# Avoid `tstate` copy + distinguish between empty/non-existent values