Before:
```
[04-14|16:35:46.445] INFO load/load_test.go:318 block generation {"t": "9.360166417s", "avg(ms)": 70, "tps": 53417.85367105188}
[04-14|16:35:46.445] INFO load/load_test.go:335 block verification {"instance": 1, "t": "8ns", "parse(ms/b)": 4.240934666666665, "parse1(ms/b)": 4.029236970149253, "parse2(ms/b)": 4.262419102941177, "verify(ms/b)": 39.5683263712121, "verify1(ms/b)": 38.22929161194029, "verify2(ms/b)": 39.142007985294114, "accept(ms/b)": 18.10627963636363, "accept1(ms/b)": 17.664332656716418, "accept2(ms/b)": 17.742920941176475, "tps": 61178.15893441096, "disk size (MB)": 0.002300262451171875}
[04-14|16:35:46.445] INFO load/load_test.go:335 block verification {"instance": 2, "t": "8ns", "parse(ms/b)": 4.2713797045454545, "parse1(ms/b)": 3.9824048656716413, "parse2(ms/b)": 4.367661691176473, "verify(ms/b)": 39.2485176590909, "verify1(ms/b)": 39.90108456716419, "verify2(ms/b)": 36.87399507352941, "accept(ms/b)": 18.525795522727268, "accept1(ms/b)": 18.65175256716417, "accept2(ms/b)": 17.584376279411767, "tps": 61049.826533749394, "disk size (MB)": 0.002300262451171875}
[04-14|16:35:46.445] INFO load/load_test.go:335 block verification {"instance": 3, "t": "8ns", "parse(ms/b)": 4.263570984848487, "parse1(ms/b)": 4.075228208955223, "parse2(ms/b)": 4.261045294117647, "verify(ms/b)": 40.78824023484848, "verify1(ms/b)": 40.40833086567164, "verify2(ms/b)": 39.36308151470587, "accept(ms/b)": 18.831947575757585, "accept1(ms/b)": 18.573730761194042, "accept2(ms/b)": 18.255545867647065, "tps": 59293.29863020369, "disk size (MB)": 0.002300262451171875}
[04-14|16:35:46.445] INFO load/load_test.go:335 block verification {"instance": 4, "t": "7ns", "parse(ms/b)": 4.238021204545455, "parse1(ms/b)": 3.979563552238806, "parse2(ms/b)": 4.305706485294118, "verify(ms/b)": 37.490303606060614, "verify1(ms/b)": 36.34277111940299, "verify2(ms/b)": 36.966976632352946, "accept(ms/b)": 18.163811863636365, "accept1(ms/b)": 17.775760582089553, "accept2(ms/b)": 17.744811867647055, "tps": 63245.01008339924, "disk size (MB)": 0.002300262451171875}
```

After (~20% slower):
```
[04-14|16:38:02.753] INFO load/load_test.go:318 block generation {"t": "9.836097125s", "avg(ms)": 74, "tps": 50833.17027534943}
[04-14|16:38:02.753] INFO load/load_test.go:335 block verification {"instance": 1, "t": "10ns", "parse(ms/b)": 4.311548310606062, "parse1(ms/b)": 3.946376268656718, "parse2(ms/b)": 4.481134808823529, "verify(ms/b)": 54.74347165909088, "verify1(ms/b)": 52.74422391044776, "verify2(ms/b)": 54.29816554411765, "accept(ms/b)": 18.416132878787874, "accept1(ms/b)": 17.906733164179098, "accept2(ms/b)": 18.105564970588237, "tps": 48894.05473656729, "disk size (MB)": 0.002300262451171875}
[04-14|16:38:02.753] INFO load/load_test.go:335 block verification {"instance": 2, "t": "10ns", "parse(ms/b)": 4.312231386363635, "parse1(ms/b)": 4.142308492537313, "parse2(ms/b)": 4.289409911764705, "verify(ms/b)": 55.51764552272727, "verify1(ms/b)": 52.93503182089551, "verify2(ms/b)": 55.61297172058825, "accept(ms/b)": 19.180657431818187, "accept1(ms/b)": 18.644698298507468, "accept2(ms/b)": 18.862529338235294, "tps": 47941.43995451436, "disk size (MB)": 0.002300262451171875}
[04-14|16:38:02.753] INFO load/load_test.go:335 block verification {"instance": 3, "t": "10ns", "parse(ms/b)": 4.242467765151517, "parse1(ms/b)": 4.016359999999998, "parse2(ms/b)": 4.278082720588236, "verify(ms/b)": 55.91182675757575, "verify1(ms/b)": 56.90566485074627, "verify2(ms/b)": 52.46590569117648, "accept(ms/b)": 18.96370833333333, "accept1(ms/b)": 19.39340549253731, "accept2(ms/b)": 17.70369605882353, "tps": 47876.31956243988, "disk size (MB)": 0.002300262451171875}
[04-14|16:38:02.753] INFO load/load_test.go:335 block verification {"instance": 4, "t": "10ns", "parse(ms/b)": 4.132102325757576, "parse1(ms/b)": 3.881102089552238, "parse2(ms/b)": 4.1971127500000005, "verify(ms/b)": 57.248463742424256, "verify1(ms/b)": 57.205584014925385, "verify2(ms/b)": 54.76504536764703, "accept(ms/b)": 19.401820378787882, "accept1(ms/b)": 19.838218925373138, "accept2(ms/b)": 18.115876794117646, "tps": 46889.90947754896, "disk size (MB)": 0.002300262451171875}
```
_Current theory is that de-deupping changes before they get to the MerkleDB improves performance_