@patrick-ogrady 2023-05-05
Also curious how many "dead bytes" we end up allocating because we have to use the max size of any state key.

It still may be more efficient but would be something to think about.